requirements.txt
-e git+http://github.com/openaustralia/scraperwiki-python.git@morph_defaults#egg=scraperwiki

bleach
beautifulsoup4
lxml
python-dateutil
regex
requests
